[Change] Full Cleanup - Manual Scan

+++ src/main/java/me/deftware/mixin/mixins/MixinRenderLivingBase.java
@@ -45,7 +45,7 @@ public abstract class MixinRenderLivingBase<T extends LivingEntity> {
                 flip = names.equals(s);
             }
             if (flip) {
-                matrixStack_1.translate(0.0D, (double)(livingEntity_1.getHeight() + 0.1F), 0.0D);
+                matrixStack_1.translate(0.0D, livingEntity_1.getHeight() + 0.1F, 0.0D);
                 matrixStack_1.multiply(Vector3f.POSITIVE_Z.getDegreesQuaternion(180.0F));
             }
         }
